requirements.txt
ultralytics==8.0.141
pandas==2.0.2
opencv-python==4.7.0.72
numpy==1.24.3
scipy==1.10.1
easyocr==1.7.0
filterpy==1.4.5
Pillow==9.5.0
paddleocr>=2.0.1
paddlepaddle==2.6.0
dill